+++ CHANGES.rst
@@ -12,6 +12,9 @@ Version 1.x is for PloneMeeting 4.1.x, version 2.x is for PloneMeeting 4.2.x+
   - element does not exist (`404 NotFound`).
 
   [gbastien]
+- Fixed serialization of `Meeting.committees` that failed on
+  `RichTextValue` in a `datagridfield`.
+  [gbastien]
 
 2.11 (2026-01-15)
 -----------------
@@ -30,7 +33,8 @@ Version 1.x is for PloneMeeting 4.1.x, version 2.x is for PloneMeeting 4.2.x+
 2.9 (2025-05-05)
 ----------------
 
-- SUP-43789: Fixed an issue with `utils.clean_html` when the ending tag </p> was not at the end of the content string.
+- SUP-43789: Fixed an issue with `utils.clean_html` when the ending tag `</p>`
+  was not at the end of the content string.
   [aduchene]
 
 2.8 (2024-10-16)